Refactor and Improvement on MiniMaxAI
Completely refactored MiniMax, removed redundant helper function (had me a dumb dumb moment when programming this before).
Added a populate board function so that evaluation could be changed to -1,0,1 scores. This is for exhaustive search once symmetry pruning is implemented. An updated version of a heuristic evaluation function was added, the old one did not evaluate a proper score, this new score should be more correct and accurate. This includes correcting analyze_line_with_traits(), as it was only checking for features with 1's not also with 0's. This heuristic is used if a score is needed for a non-terminal state.

MiniMaxAI contains the refactored code from before, this still doesn't play optimally, I believe the evaluation method is what's causing this suboptimal play, could also be that the scores aren't being propagated up through the tree properly.

MiniMaxAI_Sum uses summed scores to choose branches with better odds of winning, this still doesn't play 100% optimally however it does seem to work waaayyy better on average.

Potential bug: I noticed the game reported a stalemate however the final board configuration still had two empty places, there may be a piece tracking bug somewhere in this code, I have not been able to replicate it, but the main game loop has been updated to continue if no one has won and if pieces or spots remain. If this errors out we know we have a bug somewhere.